Fix int overflows in the ASCII fast reader c parser

+++ astropy/io/ascii/src/tokenizer.h
@@ -10,6 +10,7 @@
 #include <math.h>
 #include <float.h>
 #include <ctype.h>
+#include <sys/types.h>
 
 #ifdef _MSC_VER
     #define inline __inline
@@ -54,8 +55,8 @@ typedef enum
 typedef struct
 {
     char *source;          // single string containing all of the input
-    int source_len;        // length of the input
-    int source_pos;        // current index in source for tokenization
+    off_t source_len;      // length of the input
+    off_t source_pos;      // current index in source for tokenization
     char delimiter;        // delimiter character
     char comment;          // comment character
     char quotechar;        // quote character
@@ -105,7 +106,7 @@ double xstrtod(const char *str, char **endptr, char decimal,
 void start_iteration(tokenizer_t *self, int col);
 char *next_field(tokenizer_t *self, int *size);
 long file_len(FILE *fhandle);
-char *get_line(char *ptr, int *len, int map_len);
+char *get_line(char *ptr, size_t *len, size_t map_len);
 void reset_comments(tokenizer_t *self);
 
 #endif
